Fix backwards compatibility broken in #91
